[Reporting] Optional message should be sticky #342, Add event max length
